+++ src/main/java/com/pigmy/app/model/response/SearchTransactionResponse.java
@@ -12,7 +12,7 @@ public class SearchTransactionResponse {
 
     private Integer accountNumber;
     private String customerName;
-    private Double collectedAmount;
+    private long collectedAmount;
     private String schemeName;
     private String status;
     private  String agentName;
